Performance and bug fixes for CAN frame capture
- Lock-free SPSC ring buffer with memory barriers (replaces mutex per frame)
- LED throttling: flash at most every 100ms instead of per frame
- USB flush batching: 8 frames or 2ms max latency
- Parser stuck bug: add DISCARD_UNTIL_ETX state for oversized payloads
- Fix memory barrier placement in consumer (was causing ID 0x000 ghost frames)
- IRQ mode: use shared mutex instead of separate spinlock (fixes SPI corruption)
- Delete dead code mcp2515.c (1037 lines)
- Add shared mcp2515_read_rxb() helper
- Simplify mcp2515_receive() to delegate to receive_all()

+++ firmware/src/usb/usb_comm.h
@@ -41,8 +41,8 @@ uint32_t usb_get_rx_count(void);
 uint32_t usb_get_tx_to_host_count(void);
 
 // Debug: buffer state
-uint16_t usb_get_buffer_head(void);
-uint16_t usb_get_buffer_tail(void);
+uint32_t usb_get_buffer_head(void);
+uint32_t usb_get_buffer_tail(void);
 
 // Performance monitoring
 void usb_get_perf_stats(perf_stats_t *stats);
